pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "foresite-analytics"
version = "0.1.0"
description = "ForeSite Analytics — HR recruitment analytics agent for Canadian CMAs"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "boto3>=1.34.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "mangum>=0.17.0",
    "uvicorn>=0.29.0",
    "opensearch-py>=2.6.0",
    "requests-aws4auth>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
